swift 6

+++ Source/Classes/Extension/UI/UIColor+.swift
@@ -9,8 +9,8 @@ import Foundation
 
 // MARK: - 颜色
 
-public func RGBA(_ red: Int, _ green: Int, _ blue: Int, _ alpha: CGFloat = 1.0) -> UIColor {
-    return UIColor(red: CGFloat(red) / 255.0, green: CGFloat(green) / 255.0, blue: CGFloat(blue) / 255.0, alpha: alpha)
+public func RGBA(_ red: CGFloat, _ green: CGFloat, _ blue: CGFloat, _ alpha: CGFloat = 1.0) -> UIColor {
+    return UIColor(red: red / 255.0, green: green / 255.0, blue: blue / 255.0, alpha: alpha)
 }
 
 /// 16进制数字转换为颜色(0xAARRGGBB, 0xRRGGBB)
